+++ k8s-scheduler-apiary-iam.tf
@@ -4,9 +4,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  */
 
-resource "aws_iam_role" "beekeeper_k8s_role_scheduler_iam" {
+resource "aws_iam_role" "beekeeper_k8s_role_scheduler_apiary_iam" {
   count = var.instance_type == "k8s" ? 1 : 0
-  name  = "${local.instance_alias}-scheduler-${var.aws_region}"
+  name  = "${local.instance_alias}-scheduler-apiary-${var.aws_region}"
 
   assume_role_policy = <<EOF
 {
@@ -33,14 +33,14 @@ resource "aws_iam_role" "beekeeper_k8s_role_scheduler_iam" {
 EOF
 }
 
-resource "aws_iam_role_policy_attachment" "beekeeper_k8s_scheduler_sqs" {
+resource "aws_iam_role_policy_attachment" "beekeeper_k8s_scheduler_apiary_sqs" {
   count      = var.instance_type == "k8s" ? 1 : 0
-  role       = aws_iam_role.beekeeper_k8s_role_scheduler_iam[count.index].id
+  role       = aws_iam_role.beekeeper_k8s_role_scheduler_apiary_iam[count.index].id
   policy_arn = aws_iam_policy.beekeeper_sqs.arn
 }
 
-resource "aws_iam_role_policy_attachment" "beekeeper_k8s_scheduler_secrets" {
+resource "aws_iam_role_policy_attachment" "beekeeper_k8s_scheduler_apiary_secrets" {
   count      = var.instance_type == "k8s" ? 1 : 0
-  role       = aws_iam_role.beekeeper_k8s_role_scheduler_iam[count.index].id
+  role       = aws_iam_role.beekeeper_k8s_role_scheduler_apiary_iam[count.index].id
   policy_arn = aws_iam_policy.beekeeper_secrets.arn
 }